Fix: Change TXE interrupt handling to avoid race condition when flushing the last byte of the buffer. Now the interrupt self closes when the buffer is empty, instead of waiting for the next trigger to close it.
Feat: Add asynchronous gyroscope calibration to avoid blocking the main loop during calibration and allow processing of serial messages to be handled by the state machine.

Feat: Add ART accelerator enabling to the system initialization code to ensure instruction cache, data cache, and prefetch are enabled at boot time.

Fix: Adjust I and D components of the Delta PID controller to use int32_t for calculations to prevent overflow and ensure correct behavior with larger values.

+++ Core/pid/include/pid/controllers/delta_pid.h
@@ -24,7 +24,7 @@ const DeltaPid* get_delta_pwm_pid_ptr(void);
  * @brief Calculate and return the Delta PWM PID value.
  * @return The calculated Delta PWM PID value.
  */
-int16_t get_delta_pwm_pid(void);
+int32_t get_delta_pwm_pid(void);
 
 /**
  * @brief Check if an update is pending for the Delta PWM PID controller.
